+++ src/neurobrix/kernels/ops/binary_strided.py
@@ -28,7 +28,7 @@ def add_strided_nd_kernel(
     NDIM: tl.constexpr,
 ):
     """out = x + alpha * y, x and y read by their strides, out contiguous."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
@@ -59,7 +59,7 @@ def mul_strided_nd_kernel(
     NDIM: tl.constexpr,
 ):
     """out = x * y, x and y read by their strides, out contiguous."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
@@ -91,7 +91,7 @@ def sub_strided_nd_kernel(
     NDIM: tl.constexpr,
 ):
     """out = x - alpha * y, x and y read by their strides, out contiguous."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
@@ -122,7 +122,7 @@ def div_strided_nd_kernel(
     NDIM: tl.constexpr,
 ):
     """out = x / y, x and y read by their strides, out contiguous."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
@@ -153,7 +153,7 @@ def add_scalar_strided_nd_kernel(
     NDIM: tl.constexpr,
 ):
     """out = x + scalar, x read by its strides, out contiguous (add_scalar_kernel's expression)."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
@@ -180,7 +180,7 @@ def mul_scalar_strided_nd_kernel(
     NDIM: tl.constexpr,
 ):
     """out = x * scalar, x read by its strides, out contiguous (mul_scalar_kernel's expression)."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offsets = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offsets < n_elements
     remaining = offsets
