+++ src/neurobrix/kernels/ops/floor.py
@@ -11,7 +11,7 @@ import triton.language as tl
 # fp32 inputs take the same path unchanged.
 @triton.jit
 def floor_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.constexpr):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
     x = tl.load(input_ptr + offset, mask=mask)
@@ -20,7 +20,7 @@ def floor_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.const
 
 @triton.jit
 def ceil_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.constexpr):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
     x = tl.load(input_ptr + offset, mask=mask)
@@ -29,7 +29,7 @@ def ceil_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.conste
 
 @triton.jit
 def round_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.constexpr):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
     x = tl.load(input_ptr + offset, mask=mask)
@@ -42,7 +42,7 @@ def round_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.const
 
 @triton.jit
 def trunc_forward_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.constexpr):
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
     x = tl.load(input_ptr + offset, mask=mask)
