+++ src/neurobrix/kernels/ops/lerp.py
@@ -20,7 +20,7 @@ def lerp_tensor_forward_kernel(
     BLOCK_SIZE: tl.constexpr,
 ):
     """Lerp with tensor weight: numerically stable two-branch formula."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
@@ -45,7 +45,7 @@ def lerp_scalar_head_kernel(
     BLOCK_SIZE: tl.constexpr,
 ):
     """Lerp with scalar weight < 0.5: a + w*(b-a)."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
 
@@ -65,7 +65,7 @@ def lerp_scalar_tail_kernel(
     BLOCK_SIZE: tl.constexpr,
 ):
     """Lerp with scalar weight >= 0.5: b - (b-a)*(1-w)."""
-    pid = tl.program_id(0)
+    pid = tl.program_id(0).to(tl.int64)
     offset = pid.to(tl.int64) * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)   # 64-bit from the program id: the product itself wraps past 2^31 elements (register 58)
     mask = offset < n_elements
